requirements.txt
# Azure dependencies
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pyodbc==5.0.1
azure-storage-blob==12.19.0

# OpenAI for AI categorization and ChatGPT agent
openai==1.12.0

# Email parsing
email-parser==0.1.0
beautifulsoup4==4.12.3
html2text==2024.2.26

# Web scraping for IBAN lookup
requests==2.31.0
selenium==4.17.2
webdriver-manager==4.0.1

# Data processing
pandas==2.2.0
python-dateutil==2.8.2

# Configuration
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Logging and monitoring
applicationinsights==0.11.10

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4


# Turso Database (LibSQL)
libsql-client==0.3.0
certifi>=2025.10.5

# Flask API server pre ChatGPT GPT integráciu
flask==3.0.0
flask-cors==4.0.0
